fix: NW.js module is not defined fix.
Now the factory assignment for localforage is applied to
`global.modules.exports` instead of `modules.exports` if `module` is
undefined.
`module` is undefined in an NW.js environment (for whatever
reason). So now we check for module and we uses global if `module`
is undefined. Otherwise we proceed as before.

Thank @ashleyhindle for fix.

Fixes #77