test: Remove createInstance call
This is what was keeping the bind tests from working. It seems like it's enough
to just `$localForage.clear()`, and all of the tests still pass, so this seems
like a good move. The other idea would be to have separate functional test
suites for each "section" of the app, like the unit tests are broken up. If this
test suite grows much more, it might be worth it at that time.